Find variable features within subsets

Extension of Seurat::FindVariableFeatures that finds variable features within each subset of the data.

FindVariableFeatures_split(
  seurat,
  split.by,
  nfeatures = 4000,
  count_overall = FALSE,
  return_nested = FALSE,
  nfeatures_max = 2000,
  batches = 10
)

Arguments

seurat

Seurat object.

split.by

Metadata variable to split data into subsets.

nfeatures

The number of variable genes to select.

count_overall

Whether nfeatures should be interpreted as the number of variable genes overall (TRUE) or per subset (FALSE).

return_nested

Return a named list of variable genes per subset (TRUE) or a flattened list (FALSE).

nfeatures_max

The max number of features to return.

batches

Number of batches. Features are prioritised according to how many variable gene subsets they appear in.

Examples

if (FALSE) { # \dontrun{
var_features <- FindVariableFeatures_split(seurat=scNLP::pseudo_seurat, split.by = 'batch')
} # }